refactor: convert path into a class
